edited small mistake in parse_out

+++ srcs/get_values.c
@@ -6,7 +6,7 @@
 /*   By: thuynh <thuynh@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/04 21:05:12 by thuynh            #+#    #+#             */
-/*   Updated: 2017/08/17 11:09:08 by thuynh           ###   ########.fr       */
+/*   Updated: 2017/08/17 12:38:23 by thuynh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,6 +33,6 @@ void	get_values(t_print **list, va_list ap)
 	{
 		VAL_DBL = va_arg(ap, double);
 		VAL_DBL < 0 ? NEG++ : 0;
-		VAL_DBL = NEG ? (double)-VAL_DBL : (double)VAL_DBL;
+		VAL_DBL = NEG ? -VAL_DBL : VAL_DBL;
 	}
 }
